aula_10/react-cli/pyproject.toml
[project]
name = "react-cli"
version = "0.1.0"
description = ""
authors = [
    {name = "Ian Vieira Salgado",email = "vieira4ev@gmail.com"}
]
readme = "README.md"
requires-python = "~3.13"
dependencies = [
    "langfuse (>=3.9.3,<4.0.0)",
    "langchain (>=1.0.5,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "rich (>=14.2.0,<15.0.0)",
    "windows-curses (>=2.4.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psutil (>=7.1.3,<8.0.0)",
    "gputil (>=1.4.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "react_cli", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
